+++ vox/Source/vox/src/Box2DDebugDraw.cpp
@@ -21,7 +21,7 @@ void Box2DDebugDraw::DrawPolygon(const b2Vec2* vertices, int32 vertexCount, cons
 	sprite.setShader(&shader, true);
 	sprite.mesh->vertices.clear();
 	sprite.mesh->polygonalDrawMode = GL_POLYGON;
-	for(unsigned long int i = 0; i < vertexCount; i++) {
+	for(int32 i = 0; i < vertexCount; i++) {
 		sprite.mesh->pushVert(Vertex(vertices[i].x, vertices[i].y, 0.0001f, color.r, color.g, color.b, 1.f));
 	}
 	glLineWidth(2);
@@ -35,7 +35,7 @@ void Box2DDebugDraw::DrawSolidPolygon(const b2Vec2 * vertices, int32 vertexCount
 	sprite.setShader(&shader, true);
 	sprite.mesh->vertices.clear();
 	sprite.mesh->polygonalDrawMode = GL_POLYGON;
-	for(unsigned long int i = 0; i < vertexCount; i++) {
+	for(int32 i = 0; i < vertexCount; i++) {
 		sprite.mesh->pushVert(Vertex(vertices[i].x, vertices[i].y, 0.0001f, color.r, color.g, color.b, 1.f));
 	}
 	glLineWidth(2);
@@ -49,9 +49,9 @@ void Box2DDebugDraw::DrawCircle(const b2Vec2& center, float32 radius, const b2Co
 	sprite.setShader(&shader, true);
 	sprite.mesh->vertices.clear();
 	sprite.mesh->polygonalDrawMode = GL_POLYGON;
-	for (int i = 0; i < 30; ++i){
+	for (int32 i = 0; i < 30; ++i){
 		float percent = (i / static_cast<float>(30-1));
-		float rad = percent * 2 * 3.14159;
+		float rad = percent * 2 * 3.14159f;
 		float x = center.x + radius * cos(rad);
 		float y = center.y + radius * sin(rad);
 		sprite.mesh->pushVert(Vertex(x, y, 0.0001f, color.r, color.g, color.b, 1.f));
@@ -67,9 +67,9 @@ void Box2DDebugDraw::DrawSolidCircle(const b2Vec2& center, float32 radius, const
 	sprite.setShader(&shader, true);
 	sprite.mesh->vertices.clear();
 	sprite.mesh->polygonalDrawMode = GL_POLYGON;
-	for (int i = 0; i < 30; ++i){
+	for (int32 i = 0; i < 30; ++i){
 		float percent = (i / static_cast<float>(30-1));
-		float rad = percent * 2 * 3.14159;
+		float rad = percent * 2 * 3.14159f;
 		float x = center.x + radius * cos(rad);
 		float y = center.y + radius * sin(rad);
 		sprite.mesh->pushVert(Vertex(x, y, 0.0001f, color.r, color.g, color.b, 1.f));
